microservices/coffee_creation/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bidict==0.23.1
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorzero==2.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.1
gpiozero==2.0.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.25.1
orjson==3.10.0
packaging==24.0
pip==23.0.1
pluggy==1.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pytest==7.4.3
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
PyYAML==6.0.1
RPi.GPIO==0.7.1
setuptools==66.1.1
simple-websocket==1.0.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
ujson==5.9.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wsproto==1.2.0